pyproject.toml
[project]
name = "wms-incident-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.86.0",
    "boto3>=1.42.75",
    "bs4>=0.0.2",
    "chromadb>=1.5.5",
    "deepagents>=0.4.12",
    "faker>=40.11.1",
    "fastapi>=0.135.2",
    "fastmcp>=3.1.1",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-neo4j>=0.7.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langgraph-supervisor>=0.0.31",
    "pandas>=3.0.2",
    "prettyprinter>=0.18.0",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "sqlglot>=30.0.3",
    "sse-starlette>=3.3.3",
    "structlog>=25.5.0",
    "tabulate>=0.10.0",
    "tenacity>=9.1.4",
    "trustcall>=0.0.39",
    "uvicorn[standard]>=0.42.0",
]

[tool.setuptools.packages.find]
where = ["src"]
